+++ R/applications.R
@@ -121,6 +121,24 @@ applications <- function(account = NULL, server = NULL) {
   return(res)
 }
 
+# Use the API to filter applications by name and error when it does not exist.
+getAppByName <- function(client, accountInfo, name, error_call = caller_env()) {
+  # NOTE: returns a list with 0 or 1 elements
+  app <- client$listApplications(accountInfo$accountId, filters = list(name = name))
+  if (length(app)) {
+    return(app[[1]])
+  }
+  cli::cli_abort(
+    c(
+      "No application found",
+      i = "Specify the application directory, name, and/or associated account."
+    ),
+    call = error_call,
+    class = "rsconnect_app_not_found"
+  )
+}
+
+# Use the API to list all applications then filter the results client-side.
 resolveApplication <- function(accountDetails, appName) {
   client <- clientForAccount(accountDetails)
   apps <- client$listApplications(accountDetails$accountId)
@@ -145,7 +163,7 @@ getApplication <- function(account, server, appId) {
 }
 
 stopWithApplicationNotFound <- function(appName) {
-  stop(paste("No application named '", appName, "' is currently deployed",
+  stop(paste("No application named '", appName, "' is currently deployed.",
              sep = ""), call. = FALSE)
 }
 
@@ -227,9 +245,6 @@ showLogs <- function(appPath = getwd(), appFile = NULL, appName = NULL,
   accountDetails <- accountInfo(deployment$account, deployment$server)
   client <- clientForAccount(accountDetails)
   application <- getAppByName(client, accountDetails, deployment$name)
-  if (is.null(application))
-    stop("No application found. Specify the application's directory, name, ",
-         "and/or associated account.")
 
   if (streaming) {
     # streaming; poll for the entries directly